all: consolidate errors
